# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/account/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success\":true,\"data\":{},\"msg\":null}",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/account/register": {
            "post": {
                "description": "注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "用户注册模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success\":true,\"data\":{},\"msg\":null}",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/user/add": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toke:格式如Bearer 31a165baebe6dec616b1f8f3207b4273",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "pwd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "图像",
                        "name": "photo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success\":true,\"data\":{},\"msg\":null}",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/user/changePwd": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "修改个人密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Toke:格式如Bearer 31a165baebe6dec616b1f8f3207b4273",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新密码",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChangePwdDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success\":true,\"data\":{},\"msg\":null}",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ChangePwdDTO": {
            "type": "object",
            "required": [
                "pwd"
            ],
            "properties": {
                "pwd": {
                    "type": "string"
                }
            }
        },
        "dto.LoginDTO": {
            "type": "object",
            "required": [
                "pwd",
                "userName"
            ],
            "properties": {
                "pwd": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "dto.UserDTO": {
            "type": "object",
            "required": [
                "pwd",
                "userName"
            ],
            "properties": {
                "photo": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}